requirements.txt
librosa>=0.10.0
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
joblib>=1.3
tqdm>=4.65
Pillow>=10.0
tensorflow>=2.14        # or torch + torchaudio for PyTorch alternative
kaggle>=1.5             # for dataset download
streamlit>=1.32
librosa>=0.10